+++ CHANGELOG.md
@@ -3,10 +3,14 @@
 - Click [here](https://r35007.github.io/Mock-Server/) for mock server documentation.
 - Click [here](https://github.com/R35007/Mock-Server/blob/main/CHANGELOG.md) to see more changes reg the `@r35007/mock-server` package
 
+## v21.0.0
+
+- Added - `Serve Static Files` - Right click on any file or folder in and select this command to serve the static files.
+
 ## v20.1.0
 
 - Updated - `@35007/mock-server@18.1.1` node package.
-- Added - `MockServer: Start Server in Terminal` - THis command starts the mock server in terminal
+- Added - `MockServer: Start Server in Terminal` - This command starts the mock server in terminal
 
 ## v20.0.0